minor corrections

+++ app/src/main/java/io/gauravtak/arithmetic_calculator_with_compose/DoubleExtensions.kt
@@ -4,9 +4,9 @@ import kotlin.math.ceil
 import kotlin.math.floor
 
 fun Double.ridZero() : String {
-    val intValue = this.toInt()
-    val floorValue = floor(this).toInt()
-    val ceilValue = ceil(this).toInt()
+    val intValue = this.toLong()
+    val floorValue = floor(this).toLong()
+    val ceilValue = ceil(this).toLong()
     if (floorValue == ceilValue && intValue == floorValue) {
         return intValue.toString()
     }
